+++ src/lib_rules.ml
@@ -378,10 +378,9 @@ module Gen (P : sig val sctx : Super_context.t end) = struct
         Dep_graph.top_closed_implementations dep_graphs.impl modules
         >>^ fun modules -> modules @ wrapped_compat
       | Some (vlib_dep_graphs : Dep_graph.Ml_kind.t) ->
-        Dep_graph.top_closed_multi_implementations
-          [ vlib_dep_graphs.impl
-          ; dep_graphs.impl
-          ]
+        Dep_graph.top_closed_implementations_for_vlib_impl
+          ~vlib:vlib_dep_graphs.impl
+          ~impl:dep_graphs.impl
           modules
     in